export expandRootIds
